requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-environ==0.10.0
Pillow==9.4.0
django-import-export==3.1.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.1
gunicorn==20.1.0